Move with_spark to dataset base
