+++ client/verta/verta/dataset/_dataset.py
@@ -4,10 +4,14 @@ from __future__ import print_function
 
 import abc
 import copy
+import functools
+import hashlib
 import os
 import pathlib2
 
 from .._protos.public.modeldb.versioning import Dataset_pb2 as _DatasetService
+
+from ..external import six
 
 from .._internal_utils import (
     _file_utils,
@@ -250,6 +254,50 @@ class _Dataset(blob.Blob):
             raise KeyError("no components found for path {}".format(component_path))
 
         return (components_to_download, os.path.abspath(downloaded_to_path))
+
+    @classmethod
+    def with_spark(cls, sc, paths):
+        """
+        Creates an HDFSPath blob with a SparkContext instance.
+
+        Parameters
+        ----------
+        sc : pyspark.SparkContext
+            SparkContext instance.
+        paths : list of strs
+            List of paths to binary input data file(s) from HDFS.
+
+        Returns
+        -------
+        :class:`~verta.dataset.HDFSPath`
+            HDFSPath blob capturing the metadata of the binary files.
+
+        """
+        if isinstance(paths, six.string_types):
+            paths = [paths]
+
+        rdds = list(map(sc.binaryFiles, paths))
+        rdd = functools.reduce(lambda a,b: a.union(b), rdds)
+
+        def get_component(entry):
+            filepath, content = entry
+            h = hashlib.md5(content).hexdigest()
+            return Component(
+                path=filepath,
+                size=len(content),
+                # last_modified=metadata['modificationTime'], # handle timezone?
+                md5=hashlib.md5(content).hexdigest(),
+            )
+
+        result = rdd.map(get_component)
+        result = result.collect()
+        obj = cls(None, [])
+        obj._components_map.update({
+            component.path: component
+            for component
+            in result
+        })
+        return obj
 
     @abc.abstractmethod
     def add(self, paths):
